move puzzle files into engine/ folder
