+++ main.py
@@ -1,8 +1,8 @@
 import sys
 from PySide6.QtGui import QGuiApplication
 from PySide6.QtQml import QQmlApplicationEngine, qmlRegisterType
-from puzzle_handler import PuzzleBridge
-from puzzle_setting import PuzzleSetting
+from engine.puzzle_handler import PuzzleBridge
+from engine.puzzle_setting import PuzzleSetting
 import os
 
 if __name__ == "__main__":
